Filtrado de secciones por atributos y programas especiales En el panel lateral se mostraban los cursos especiales, pero no se filtraban de acuerdo al horario en pantalla
